Makes flags POSIX compliant
