+++ README.md
@@ -24,20 +24,22 @@ go build
 ## Usage
 
 ```None
-:~$ ./sourcemapper
-Usage of ./sourcemapper:
-  -header value
-    	A header to send with the request, similar to curl's -H. Can be set multiple times, EG: "./sourcemapper --header "Cookie: session=bar" --header "Authorization: blerp"
-  -help
-    	Show help
-  -insecure
-    	Ignore invalid TLS certificates
-  -output string
-    	Source file output directory - REQUIRED
-  -proxy string
-    	Proxy URL
-  -url string
-    	URL or path to the Sourcemap file - REQUIRED
+./sourcemapper --help
+
+A golang tool to parse a sourcemap as generated by webpack
+or the like and spit out the original JavaScript files,
+recreating the source tree based on the file paths in the sourcemap.
+
+Usage:
+  sourcemapper <URL or path> [flags]
+
+Flags:
+  -H, --header stringArray   Send headers with the request, similar to curl's -H. Can be set multiple times. For example:
+                             ./sourcemapper --header "Cookie: session=bar" --header "Authorization: blerp"
+                             
+  -h, --help                 help for sourcemapper
+  -k, --insecure             Ignore invalid TLS certificates
+  -o, --output string        Source file output directory
 ```
 
 Sourcemapper will download or read the map file at `url`, and then spit the sources out into the directory defined by `output`.
