+++ go.mod
@@ -1,3 +1,5 @@
 module github.com/denandz/sourcemapper
 
 go 1.16
+
+require github.com/spf13/cobra v1.6.1
